after reverting the previous commit...
it turns out that the clink map needs special treatment. this commit
restores everything that was reverted in the last one except for the
changes to the clink map. note to self, it looks like it is trying to be
an adjacency matrix for an undirected graph, so we may be able to cut
the storage requirement in half...