code cleanup:
remove "possible_fake_data" map, it was identical to the nreads_FR map

store LibraryInfo in a vector instead of a map. provide a map from lib
names -> vector indices. this is in preparation to refer to these by
index rather than by name throughout the code

instead of copying map keys to vectors to iterate over in
buildConnection, just iterate over the map making sure not to invalidate
iterators while erasing edges from the graph.