Pulled out the grpah code into its own class
I was thinking this was going to help me prune vertices a little
bit more aggressively, but after writing it, I realized that it is more
time/space efficient to do it differently. In any case, if we want to
fiddle with the implementation of the graph, it's off in its own class
now.